update8

+++ UserInterface.h
@@ -47,7 +47,7 @@ int UserInterface ::count(const string& input)
 {
 	int Spaces = 0;
     
-	for ( int i = 0; i < input.size(); ++i)
+	for ( size_t i = 0; i < input.size(); ++i)
 		if (input[i] == ' ') ++Spaces;
     
 	return Spaces;
